agents/document/requirements.txt
langgraph>=0.2.0
python-dotenv>=1.0.1
reportlab>=4.2.0
markdown>=3.6
beautifulsoup4>=4.12.3
